backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
agno>=1.4.2
openai>=1.74.0
redis>=5.0.1
psycopg[binary]>=3.2.9
playwright>=1.40.0
sqlalchemy>=2.0.23
asyncpg>=0.29.1
pgvector>=0.2.4
websockets>=12.0